Fix dependency cruiser issue
